feat(core): add exponential backoff for graph recomputation
